Refactor mkdocs.yml to remove toc.integrate feature
The 'toc.integrate' feature was removed from the mkdocs.yml configuration file. This change was made to simplify the site navigation and improve the user experience. The 'toc.integrate' feature allowed for integrating table of contents into the content area, but it was decided that it was not necessary for the current use case. This change does not have any significant consequences other than removing a less useful feature. The commit addresses