Create a CNAME file with 'docs.katson.com' as the content
The commit adds a new file named 'CNAME' with the content 'docs.katson.com'. This change is necessary to configure the domain for the documentation site hosted at 'katson.com'. The 'CNAME' file specifies the primary domain for the site, allowing users to access the documentation using the URL 'docs.katson.com'. There are no associated issues for this change.

Note: The generated commit message assumes that the purpose of adding the 'CNAME' file is to configure the domain for the documentation site. Please modify the commit message accordingly if this assumption is incorrect.